Add separate iso7816_strings abstraction
The intention is that this abstraction will stringify data fields that
may be relevant to multiple areas of ISO7816, as well as other
abstractions such as EMV. For example, country code, life cycle cycle
status byte, and other COMPACT-TLV field values.

Currently, fields that are specific to a particular domain, like ATR
interface bytes, or indicate domain specific field types, like
COMPACT-TLV tags, are not included. But that may change in future.