Implement basic EMV Terminal Transport Layer (TTL) transceive function
The intention is that this transceive function will support the mapping
of APDUs from the Terminal Application Layer (TAL) to the appropriate
TPDUs or APDUs supported by the card reader, and to process the commands
in accordance with the cases described in ISO 7816-3:2006, section 12.

For now, only card readers in APDU mode is supported. TPDU mode will be
added in future.