Implement EMV TLV list abstraction
An EMV TLV field consists of the parsed BER field as well as a next
pointer to allow it to be added to a linked list. The intention is that
EMV TLV fields would not be created manually, or outside of a list.
Instead, the intention is that EMV TLV fields will be created when their
data is pushed onto a list. Every list entry is dynamically allocated
and the list must either be cleared using emv_tlv_list_clear() when
done, or individual entries that were removed from the list must be
free'd using emv_tlv_free() when done.

Runtime checks and asserts have been added to enforce the above.