Implement EMV application list abstraction
The EMV application structure has a next pointer to allow it to be added
to a linked list. The intention is that EMV application objects are
first created and then passed to emv_app_list_push() to push onto a list
without duplicating the object. The list must be cleared using
emv_app_list_clear() when done, or individual entries that were removed
from the list must be free'd using emv_app_free() when done.

Runtime checks and asserts have been added to enforce the above.