Move stringify functions to separate emv_strings.{h,c} files
This makes it easier to completely omit the stringify functions from
embedded builds that may be short on code space. The intention is that
bit definitions for EMV fields remain in emv_fields.h but that stringify
functions that are not necessary for EMV kernel processing by added to
emv_strings.{h,c} instead.