Implement initial EMV Data Object List (DOL) processing
This implementation assumes the length required by the DOL entry and the
length provided by the TLV field are the same. EMV 4.3 Book 3, 5.4
indicates that the format of the TLV field must be known to determine
implement trimming or padding to match the required DOL entry length,
but currently this would require all of emv_strings.c, which is not
suitable for embedded builds.