Add CMake module for iso-codes
The module file makes use of the pkgconfig file to obtain the version
and domains variable. It then validates the location of the JSON data
files and makes that available as IsoCodes_JSON_PATH. The intention is
that the latter can be used to generate preprocessor definitions for
applications to find the JSON data files. The IsoCodes_JSON_PATH is also
listed as the first item of REQUIRED_VARS such that it is displayed
during configuration of the build.