Update Application Interchange Profile (field 82) to EMV 4.4
Note that EMV Contactless Book C-3 v2.10, Annex A.2, defines AIP byte 2
bit 8 to be "Mag-stripe mode is supported". This is the opposite of the
definition in EMV Contactless Book C-2 v2.10, Annex A.1.16 of
"Contactless EMV mode is supported". However, EMV Contactless Book C-3
v2.10, Annex A.2, also states that products supporting that
specification should set this bit to zero. The assumption is therefore
that this specification no longer uses this bit and the definition in
EMV Contactless Book C-2 v2.10 is used instead when decoding this field.