Update Github Actions and tests for libintl/gettext
The CMake check for libintl is not mandatory because libintl is only for
testing locale functionality that is not actually used by any of the
project libraries at this time. In addition, brew doesn't have universal
binaries for gettext, and there's currently no value in breaking the
universal binary build for testing functionality that isn't used by any
of the libraries at this time.