Add CMake module for json-c
Newer versions of json-c have a CMake config file while older versions,
like the one for Ubuntu 20.04, do not. This module file makes use of the
pkgconfig file to obtain the version and paths, and then defines the
json-c::json-c target to mimic that of the CMake config file in newer
version of json-c. The intention is that one can first try to find the
CMake config file and if it is not available, fallback to this module.