Stringify "CVM not performed" for CVM Results (field 9F34)
A CVM code of 0x3F is invalid for CV Rules but valid for CVM Results and
indicates "CVM not performed" according to EMV 4.4 Book 4, Annex A4,
Table 33. EMV Contactless Book C-2 v2.10, 7.5, also uses this value to
indicate the absence or failure of CVM processing when the amount is
below the CVM limit, the Application Interchange Profile indicates that
CVM is not supported, the card doesn't provide a CVM list, or CVM
processing fails.