Stringify Terminal Transaction Qualifiers (field 9F66)
Note that 9F66 is defined as TTQ with a length of 4 bytes by the Entry
Point, kernel 3 (Visa), kernel 4 (Amex), kernel 6 (Diners/Discover) and
kernel 7 (UnionPay), while kernel 3 (Mastercard) defines it as a
different 2 byte field. Therefore, this field will only be decoded as
TTQ when it has a length of 4 bytes.