Add mcc-codes as a submodule
The mcc-codes project is hosted at https://github.com/greggles/mcc-codes
and provides a public repository of Merchant Category Codes (MCC) in
various formats. It is licensed under the Unlicense licence which
waives copyright and warranty.

There are various forks of this project that provide various
improvements and corrections. In future it may be of interest to fork it
under OpenEMV to incorporate the various improvements and corrections.