Implement ISO 18245 Merchant Category Code (MCC) lookup
This implementation uses the default installation path of mcc_codes.json
unless otherwise specified to mcc_init(). Unit tests can therefore
specify the build path to mcc_init() such that they do not rely on
files that have not yet been installed.

The JSON is parsed to extract the "edited_description" field for each
MCC and make it available in a std::map for lookup.