Stringify Authorisation Response Code (field 8A)
The description strings are from ISO 8583:2021, J.2.2.2, table J.3,
which according to that document is recommended to replace the table in
the published version of ISO 8583:1987. Not all of these authorisation
response codes are relevant for EMV processing and perhaps this
implementation will be moved to a future ISO 8583 library instead.