Implement PC/SC retrieval of PICC UID
The recommended way for PC/SC to determine whether the card is
contactless is to check for the artificial ATR that PC/SC creates for
contactless cards and then to attempt retrieval of the PICC UID.

The primary purpose of this change is to reject contactless cards for
now to avoid decoding their artificial ATRs. Eventually this project
will grow to support contactless cards and implement further PC/SC
contactless functionality.