Implement PC/SC reader feature list retrieval
The pcsc_reader_has_feature() function can be used to determine whether
any specific feature is supported by the PC/SC reader while this project
only provides feature definitions for features that are likely to be of
interest to this project in future. However, none of these features are
in use by this project yet.

Also stringify these features in emv-tool.