Stringify Cryptogram Information Data (field 9F27)
The payment system-specific cryptogram bits are stringified verbatim
when non-zero because there is no indication in the specification of how
these bits should be assessed even if they are proprietary. Similarly it
is not clear whether the reason/advice codes matter if advice is not
required and therefore both scenarios are stringified to avoid
ambiguity.