Implement basic EMV highlighter for emv-viewer
QSyntaxHighlighter is designed to process one block at a time with very
little context about other blocks. This is not ideal for EMV parsing but
appears to be the only way to apply text formatting without impacting
the undo/redo stack. Therefore, this implementation processes all blocks
and reparses all of the EMV data when updating each block. This
implementation also assumes that rehighlight() is called whenever the
widget text changes because changes to later blocks may invalidate EMV
field lengths specified in earlier blocks.