Improve EmvHighlighter to emphasise tags
Tags are emphasised using the "Solarized blue" colour because it appears
to contrast well against both black and white text for light and dark
themes. This allows users to easily identify tags in the input data.

This commit also converts parseBerData() to a template function such
that it can accept a lambda function as a parameter.