Ensure that argp keys are not interpreted as short options
The argp documentation indicates that if the key parameter has a value
that is a printable ASCII value, it will also be used as the short
option. Currently enums provide the key parameter for each option and
there is no intention for these to be short options. Therefore it is
best to renumber the enums to avoid printable ASCII values.