Highlight input bytes when clicking tree item in emv-viewer
EmvTreeItem is now aware of its offset and length within the source data
such that this is available when the itemPressed signal is processed.

Selected input bytes are highlighted using the "Solarised Base00" colour
because it appears to contrast well against both black and white text
for light and dark themes.