Install library dependencies for Windows builds
Use CMake's RUNTIME_DEPENDENCY_SET feature to generate and install
library dependencies like json-c and boost for iso8859 and emv_strings.
This approach was more robust than CMake's IMPORTED_RUNTIME_ARTIFACTS
because the latter only installs the imported artifact and not the
artifact's dependencies, which is especially problematic for boost.

It is unknown whether any of this would work for non-MSYS2 builds
because the search paths may differ for other environments.