Implement BUILD_WIN_STANDALONE option for Windows standalone builds
The BUILD_WIN_STANDALONE option installs external library dependencies
as well as data dependencies such as iso-codes files. This option is
disabled by default and when not set will yield a Unix style build and
packaging that may be more useful for developers and package managers
like vcpkg. When this option is set, it will yield a standalone
installation that can be packaged into an NSIS installer.