Move cached fields to EMV processing context
This allows emv_initiate_application_processing() to populate these
fields for use by various functions. On the one hand, caching these
fields may seem like an unnecessary optimisation, but on the other hand
it means that their presence/absence becomes a matter of internal errors
instead of having to determine what the appropriate external error would
be for each function that depends on these fields.