Update Github Actions workflow for Windows release builds
The latest MinGW provided by MSYS2 and used for release builds has
changed how clock_gettime() is linked. However, during packaging the
MinGW provided by Qt is used as the runtime of the installed package.
This prevents clock_gettime() from being available at runtime.

The solution is to use the MinGW provided by Qt for the build itself to
ensure that the same toolchain is used for the build and the packaging.
As such, the Qt version must be updated to a newer version with a newer
MinGW, in this case Qt-6.8.x. It is also an opportunity to update
jurplel/install-qt-action version.

Note that MSYS2's toolchain must still be available because the
packaging tools that embed resources into Windows executables still use
that toolchain.