Implement separate decode objects checkbox for emv-viewer
The identification, decoding and rendering of ASN.1 objects was
previously governed by the same checkbox that governs EMV / ASN.1 field
decoding. Instead, every EmvTreeItem instance will now have a separate
string that is only populated when it is an object, and know when to
hide itself when it is the first member of an object.