Identify and stringify ASN.1 objects in TLV linked lists
This applies the same approach as is implemented for BER encoded TLV
data by f3028e6a1461cd60fe18dddfc21833a6486d2b10. The only difference is
that the top level TLVs are already parsed into a linked list.

When printing TLVs, this implementation will assume that any TLV with a
value string is an object of some kind. This allows the identification
of other kinds of objects in future.