Implement emv_tlv_list_push_asn1_object()
This is a helper function to encode the OID for an ASN.1 object before
pushing it onto an EMV TLV list. The typical use of this function is to
specify non-EMV information (like the merchant URL or merchant email
address) in the terminal configuration such that those can eventually be
added to the transaction record.