Stringify EMV GENERATE AC command
This implementation only provides a one-line description to
emv_capdu_get_string(). Note that the difference between the string
provided for the decoding of the C-APDU and the string provided for
GENAC during EMV processing by the TAL is intentional because the former
may be used for a general decoder as well.