Implement EMV TLV sources abstraction as well as an iterator
The intention is that this abstraction can be used in future for DOL
processing as well as for finding fields used when stringifying other
fields.

The iterator allows the search for a field to continue where the
previous search stopped. This is especially useful for finding all
instances of a field that may occur multple times in multiple lists.