Stringify ICC Public Key Certificate (field 9F46)
This implementation will try all available instances of the Issuer
Public Key Certificate (field 90) to decrypt the ICC Public Key
Certificate. It relies on the emv_decrypt_issuer_pkey() function being
able to not only decrypt the Issuer Public Key Certificate (field 90)
using a CAPK, but also retrieve and validate the issuer public key.

However, full retrieval of the ICC public key is not feasible because
the ODA hash is not available to validate the public key.