Improve library dependencies for Windows builds
Due to a change in MSYS2, Boost::locale now links statically by default,
which in turn causes CMake not to find the dynamically linked DLLs for
ICU during packaging of the release build. Setting Boost_USE_STATIC_LIBS
forces dynamic linking of Boost::locale and assumes that the DLLs for
ICU can be found at the same path of those for Boost::locale during
packaging of the release build.

Use CMake's LOCATION property to retrieve the location of dependencies
that are IMPORTED targets. This will still be problematic for static
dependencies that themselves have DLLs as dependencies but there doesn't
appear to be an elegant CMake solution to address this.

The $<PATH:GET_PARENT_PATH,...> generator requires CMake >=3.24 which is
fine because Windows release builds already assume a very up-to-date
build environment.

Exclude the C/C++ runtime from the runtime dependency set of library
targets to avoid ambiguity if multiple C/C++ toolchains are available.
Instead, let windeployqt be fully responsible for providing the C/C++
runtime.