Improve PC/SC initialisation
Let the return value of pcsc_init() indicate whether no readers were
found. This is useful because SCardListReaders() may return
SCARD_E_NO_READERS_AVAILABLE when no readers are available instead of
allowing initialisation of the context to continue for that condition to
be detected later when the readers are listed.